Use "" imports instead of <> in main fmwk header
Xcode 7.1 seems to be having issues with the framework being the same module name and existing as 2 different targets (one tvOS, and one iOS). It complains in the tvOS app that the framework includes 'non-modular' headers in the framework. Changing it to quotes for now seems to fix this. Let's revisit with Xcode 7.2