For now, immediately connect to 1st service found
Later on we want to handle some sort of choice of which available peer to connect to. This is admittedly unlikely, that you'd have 2 different people both responding to opening the app to authorize the tvOS app. But the possibility is there so we should handle it.

One solution: Connect to 1st found, but if more than one service is available, then ask user to confirm in using that service, vs. try for another.