Don't stop advertising server if connected.
This was causing way too many issues with the server going "offline" while maintaining a connection. Server wanting to restart automatically, client attempting to re-connect to the server, etc.

Now the mental model is simpler: Server is publishing for as long as it (and connections) are alive. If a server is connected to a client, it's still published on the network (it rejects other connections while connected to a client).