Add convenience initializer, with just uniqueId
This makes setup easier, as pretty much everyone would be using the device name as the display name.

This introduces UIKit dependency on both VoucherClient and VoucherServer, but thankfully it's only to determine the display name via:

[UIDevice currentDevice].name

If we support more platforms in the future, we should pre-process the code using things like #TARGET_OS_TV, etc.